Address PR feedback.
- Change FunctionProperties into a concrete class.
- Move FunctionPropertiesConfig to test since it's only used for
testing. Required chaning how PPLServiceTest and StandaloneIT
instantiate FunctionProperties bean.
