+++ core/src/main/java/org/opensearch/sql/analysis/AnalysisContext.java
@@ -12,7 +12,6 @@ import java.util.Objects;
 import lombok.Getter;
 import org.opensearch.sql.expression.NamedExpression;
 import org.opensearch.sql.expression.function.FunctionProperties;
-import org.opensearch.sql.expression.function.QueryFunctionProperties;
 
 /**
  * The context used for Analyzer.
@@ -39,7 +38,7 @@ public class AnalysisContext {
   public AnalysisContext(TypeEnvironment environment) {
     this.environment = environment;
     this.namedParseExpressions = new ArrayList<>();
-    this.functionProperties = new QueryFunctionProperties();
+    this.functionProperties = new FunctionProperties();
   }
 
   /**
