+++ core/src/main/java/org/opensearch/sql/expression/function/FunctionProperties.java
@@ -7,16 +7,48 @@ package org.opensearch.sql.expression.function;
 
 import java.io.Serializable;
 import java.time.Clock;
+import java.time.Instant;
+import java.time.ZoneId;
+import lombok.EqualsAndHashCode;
+import lombok.RequiredArgsConstructor;
 
-public interface FunctionProperties extends Serializable {
-  Clock getSystemClock();
+@RequiredArgsConstructor
+@EqualsAndHashCode
+public class FunctionProperties implements Serializable {
 
-  Clock getQueryStartClock();
+  private final Instant nowInstant;
+  private final ZoneId currentZoneId;
+
+  /**
+   * By default, use current time and current timezone.
+   */
+  public FunctionProperties() {
+    nowInstant = Instant.now();
+    currentZoneId = ZoneId.systemDefault();
+  }
+
+  /**
+   * Method to access current system clock.
+   * @return a ticking clock that tells the time.
+   */
+  public Clock getSystemClock() {
+    return Clock.system(currentZoneId);
+  }
+
+  /**
+   * Method to get time when query began execution.
+   * Clock class combines an instant Supplier and a time zone.
+   * @return a fixed clock that returns the time execution started at.
+   *
+   */
+  public Clock getQueryStartClock() {
+    return Clock.fixed(nowInstant, currentZoneId);
+  }
 
   /**
    * Use when compiling functions that do not rely on function properties.
    */
-  FunctionProperties None = new FunctionProperties() {
+  public static final FunctionProperties None = new FunctionProperties() {
     @Override
     public Clock getSystemClock() {
       throw new UnexpectedCallException();
