+++ core/src/test/java/org/opensearch/sql/expression/datetime/DateTimeTestBase.java
@@ -28,7 +28,6 @@ import org.opensearch.sql.expression.env.Environment;
 import org.opensearch.sql.expression.function.BuiltinFunctionName;
 import org.opensearch.sql.expression.function.BuiltinFunctionRepository;
 import org.opensearch.sql.expression.function.FunctionProperties;
-import org.opensearch.sql.expression.function.QueryFunctionProperties;
 
 @ExtendWith(MockitoExtension.class)
 public class DateTimeTestBase extends ExpressionTestBase {
@@ -44,7 +43,7 @@ public class DateTimeTestBase extends ExpressionTestBase {
 
   @BeforeAll
   public static void setup() {
-    functionProperties = new QueryFunctionProperties();
+    functionProperties = new FunctionProperties();
   }
 
   protected Expression nullRef = DSL.literal(ExprNullValue.of());
